chore(api): remove test-only function from production helper file
