+++ apps/codecov-api/billing/tests/test_helpers.py
@@ -1,6 +1,7 @@
 from django.test import TestCase, override_settings
 
-from billing.helpers import mock_all_plans_and_tiers, on_enterprise_plan
+from billing.helpers import on_enterprise_plan
+from billing.tests.mocks import mock_all_plans_and_tiers
 from shared.django_apps.codecov_auth.tests.factories import OwnerFactory
 from shared.plan.constants import PlanName
