+++ apps/codecov-api/codecov_auth/tests/test_admin.py
@@ -8,7 +8,7 @@ from django.test import RequestFactory, TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from billing.helpers import mock_all_plans_and_tiers
+from billing.tests.mocks import mock_all_plans_and_tiers
 from codecov.commands.exceptions import ValidationError
 from codecov_auth.admin import (
     AccountAdmin,
