+++ apps/codecov-api/graphql_api/tests/test_plan_representation.py
@@ -4,7 +4,7 @@ from django.test import TestCase
 from django.utils import timezone
 from freezegun import freeze_time
 
-from billing.helpers import mock_all_plans_and_tiers
+from billing.tests.mocks import mock_all_plans_and_tiers
 from shared.django_apps.core.tests.factories import OwnerFactory
 from shared.plan.constants import DEFAULT_FREE_PLAN, PlanName, TrialStatus
